imgur/requirements.txt
Pillow==7.1.2
PyYAML==5.3.1
imgurpython==1.1.7
interruptingcow==0.8
progress==1.5
requests==2.23.0